requirements.txt
Flask~=3.1.1
kafka-python>=2.0.2
pandas~=2.2.3
